requirements.txt
Django==6.0
Pillow==12.0.0
django-cors-headers==4.9.0
django-filter==24.3
djangorestframework==3.16.1
django-jazzmin==3.0.1
razorpay==1.4.2
